backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
chardet==3.0.4
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pillow==12.1.1
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.5
tzdata==2025.3
